### 2016-12-08 e46dd81
Changes filter_length argument to nsamples_before, adds nsamples_after as a kwarg to filtertools
