### 2016-12-08 96eeeb6
Adds acausal functionality to filtertools.{revcorr,linear_response}. Renames filtertools.linear_prediction -> filtertools.linear_response. Adds acausal functionality to stimulustools.slicestim. Improves docstrings for most of filtertools module.
