+++ pyret/filtertools.py
@@ -15,36 +15,43 @@ from pyret.utils import flat2d
 
 __all__ = ['ste', 'sta', 'stc', 'lowranksta', 'decompose',
            'filterpeak', 'smooth', 'cutout', 'resample', 'flat2d',
-           'get_ellipse', 'get_contours', 'get_regionprops',
-           'normalize_spatial', 'linear_prediction', 'revcorr']
+           'get_ellipse', 'get_regionprops', 'normalize_spatial', 
+           'linear_response', 'revcorr']
 
 
 def ste(time, stimulus, spikes, nsamples_before, nsamples_after=0):
     """
-    Constructs an iterator over spike-triggered stimuli
+    Constructs an iterator over spike-triggered stimuli.
 
     Parameters
     ----------
     time : ndarray
-        The time array corresponding to the stimulus
+        The time array corresponding to the stimulus.
 
     stimulus : ndarray
-        A spatiotemporal or temporal stimulus array
-        (where time is the first dimension)
+        A spatiotemporal or temporal stimulus array, where time is the 
+        first dimension.
 
     spikes : iterable
-        A list or ndarray of spike times
+        A list or ndarray of spike times.
 
     nsamples_before : int
-        Number of samples to include in the STE before the spike
+        Number of samples to include in the STE before the spike.
 
     nsamples_after : int
-        Number of samples to include in the STE after the spike (default: 0)
+        Number of samples to include in the STE after the spike,
+        which defaults to 0.
 
     Returns
     -------
     ste : generator
-        A generator that yields samples from the spike-triggered ensemble
+        A generator that yields samples from the spike-triggered ensemble.
+
+    Notes
+    -----
+    The spike-triggered ensemble (STE) is the set of all stimuli immediately
+    surrounding a spike. If the full stimulus distribution is p(s), the STE 
+    is p(s | spike).
 
     """
     nb, na = nsamples_before, nsamples_after
@@ -64,9 +71,7 @@ def ste(time, stimulus, spikes, nsamples_before, nsamples_after=0):
 
 def sta(time, stimulus, spikes, nsamples_before, nsamples_after=0):
     """
-    Compute a spike-triggered average
-
-    sta, tax = sta(time, stimulus, spikes, filter_length)
+    Compute a spike-triggered average.
 
     Parameters
     ----------
@@ -89,10 +94,38 @@ def sta(time, stimulus, spikes, nsamples_before, nsamples_after=0):
     Returns
     -------
     sta : ndarray
-        The spatiotemporal spike-triggered average (RF)
+        The spatiotemporal spike-triggered average (RF). Note that time
+        increases with increasing array index, i.e. time of the spike is
+        at the index for which ``tax == 0``.
 
     tax : ndarray
-        A time axis corresponding to the STA
+        A time axis corresponding to the STA, giving the time relative
+        to the spike.
+
+    Notes
+    -----
+    The spike-triggered average (STA) is the averaged stimulus feature
+    conditioned on the presence of a spike. This is widely-used method
+    for estimating a neuron's receptive field, and captures the average
+    stimulus feature to which the neuron responds.
+
+    Formally, the STA is defined as the function [1]:
+
+    .. math::
+        C(\\tau) = \\frac{1}{N} \\sum_{i=1}^{N} s(t_i - \\tau)
+
+    where :math:`\\tau` is time preceding the spike, and :math:`t_i` is
+    the time of the ith spike.
+
+    The STA is often used to estimate a linear filter which captures
+    a neuron's responses. If the stimulus is uncorrelated (spherical),
+    the STA is unbiased and proportional to the time-reverse of the
+    linear filter.
+
+    References
+    ----------
+    [1] Dayan, P. and L.F. Abbott. Theoretical Neuroscience: Computational
[…]
