### 2016-12-08 929d986
Adds tests for the acausal functionality included in previous commit. Fixes off-by-one errors in that code. Updates some docs with more information about the nbefore/nafter concept.
