### 2016-12-08 827838c
Adds note to quickstart about relationship between STA, RF, and linear filter. Clarifies docstrings a bit more about which is accepted in which functions.
