+++ pyret/filtertools.py
@@ -59,11 +59,11 @@ def ste(time, stimulus, spikes, nsamples_before, nsamples_after=0):
     # Bin spikes
     (hist, bins) = np.histogram(spikes, time)
 
-    # Get indices of non-zero firing, truncating spikes earlier
-    # than `filterlength` frames
+    # Get indices of non-zero firing, truncating spikes before nsamples_before
+    # or after nsamples_after
     slices = (stimulus[(idx - nb):(idx + na), ...].astype('float64')
               for idx in np.where(hist > 0)[0]
-              if (idx > nsamples_before and (idx + nsamples_after) < len(stimulus)))
+              if (idx > nb and (idx + na) < len(stimulus)))
 
     # return the iterator
     return slices
@@ -94,7 +94,7 @@ def sta(time, stimulus, spikes, nsamples_before, nsamples_after=0):
     Returns
     -------
     sta : ndarray
-        The spatiotemporal spike-triggered average (RF). Note that time
+        The spatiotemporal spike-triggered average. Note that time
         increases with increasing array index, i.e. time of the spike is
         at the index for which ``tax == 0``.
 
@@ -213,48 +213,51 @@ def stc(time, stimulus, spikes, nsamples_before, nsamples_after=0):
     return stc
 
 
-def lowranksta(f_orig, k=10):
+def lowranksta(sta_orig, k=10):
     """
-    Constructs a rank-k approximation to the given spatiotemporal filter.
-    This is useful for computing a spatial and temporal kernel of an STA,
-    or for denoising.
+    Constructs a rank-k approximation to the given spatiotemporal STA.
+    This is useful for estimating a spatial and temporal kernel for an
+    STA or for denoising.
 
     Parameters
     ----------
-    f : array_like
-        3-D filter to be separated, shaped as (time, space, space).
+    sta_orig : array_like
+        3D STA to be separated, shaped as ``(time, space, space)``.
 
     k : int
-        Number of components to keep (rank of the filter).
+        Number of components to keep (rank of the reduced STA).
 
     Returns
     -------
-    fk : array_like
-        The rank-k estimate of the original filter.
+    sk : array_like
+        The rank-k estimate of the original STA.
 
     u : array_like
         The top ``k`` temporal components (each column is a component).
 
     s : array_like
-        the top ``k`` singular values.
+        The top ``k`` singular values.
 
     v : array_like
-        the top ``k`` spatial components (each row is a component). These
+        The top ``k`` spatial components (each row is a component). These
         components have all spatial dimensions collapsed to one.
 
     Notes
     -----
-    This method requires that the linear filter be 3D. To decompose a
-    linear filter into a temporal and 1-dimensional spatial filter, simply
-    promote the filter to 3D before calling this method.
+    This method requires that the STA be 3D. To decompose a STA into a 
+    temporal and 1-dimensional spatial component, simply promote the STA 
+    to 3D before calling this method.
+
+    Despite the name this method accepts both an STA or a linear filter.
+    The components estimated for one will be flipped versions of the other.
 
     """
 
-    # work with a copy of the filter (prevents corrupting the input)
-    f = f_orig.copy() - f_orig.mean()
+    # work with a copy of the STA (prevents corrupting the input)
+    f = sta_orig.copy() - sta_orig.mean()
 
-    # Compute the SVD of the full filter
-    assert f.ndim >= 2, "Filter must be at least 2-D"
+    # Compute the SVD of the full STA
+    assert f.ndim >= 2, "STA must be at least 2-D"
[…]
