### 2016-12-13 #137 @whedon
[REVIEW]: Pyret: A Python package for analysis of neurophysiology data
**Submitting author:** @nirum (<a href="http://orcid.org/0000-0002-3946-4705">Niru Maheswaranathan</a>)
**Repository:** <a href="https://github.com/baccuslab/pyret" target ="_blank">https://github.com/baccuslab/pyret</a>
**Version:** v0.5.0
