### 2016-12-13 baccuslab/pyret#83 @kdesimone
Support for Python 2.7?
I guess its OK to be compliant with 3 only, but you should probably make mention of this somewhere  in the documentation or the repository so that users won't be baffeled when trying to install on Python 2.7.

$ python setup.py install
Traceback (most recent call last):
  File "setup.py", line 2, in <module>
    import pyret
  File "/Users/kevin/pyret/pyret/__init__.py", line 29, in <module>
    from pyret import *
  File "/Users/kevin/pyret/pyret/nonlinearities.py", line 9, in <module>
    from itertools import zip_longest
ImportError: cannot import name zip_longest
