### 2016-12-13 baccuslab/pyret#83 @bnaecker
Several of key components of the 'filtertools' module rely heavily on Python 3's better support for lazy evaluation and iteration. There are several other smaller differences that we'd have to account for, but this large one I think means we don't plan to support 2.7 at all.

But you are correct that the documentation should much more clearly indicate that the package is 3 only. We will make updates to the docstrings and quickstart on ReadTheDocs to make this very clear.
