### 2016-12-13 baccuslab/pyret#83 @kdesimone
OK fair enough, although the benefits of widening your net to attract more users might outweigh the performance gains derive from lazy iteration and evaluation that comes with 3+ but not 2.7.
