### 2016-12-14 baccuslab/pyret#83 @bnaecker
Yes, we've thought about the issue of the size of the user base. We'all look more carefully into supporting 2.7. Both Niru (the other developer) and I are traveling out of the country now, and will look into it in a few days when we return. Also, please excuse any late or short replies for the same reason.
