### 2016-12-14 baccuslab/pyret#83 @kdesimone
As for the JOSS submission, I don't see any problem with building applications with 3+ in mind, but 2.7 compatibility is something to think about.
