### 2016-12-16 baccuslab/pyret#85 @kdesimone
JOSS review: pip requirements not installing?
Hi,

I'm trying to install pyret on a clean conda environment running Python 3.5.  Trying to install via pip, as in `pip install pyret`,  I get an import error: `ImportError: No module named 'sklearn'`.  I see that you do include install_requires in your setup, but there must be something else off in the setup.

I know [this](https://github.com/kdesimone/popeye/blob/master/setup.py) one works.
