### 2016-12-16 baccuslab/pyret#85 @kdesimone
Also pursuant to the installs—the pyret demo requires h5py but this isn't included in in the install_requirements and there are no errors related to h5py in the piecemeal install of pyret ...
