### 2016-12-16 baccuslab/pyret#86 @kdesimone
JOSS review: demo broken, related to h5py?
I think the demo assumes that you `cd` into pyret/docs in order load the `tutorial-data.h5` dataset.  What happens if  you install via `pip` and not from source?

Also, once I've navigated to pyret/docs, I get this error:

`AttributeError                            Traceback (most recent call last)`
`<ipython-input-6-ec7d48997ae9> in <module>()`
`      7 spikes = data_file['spike-times']  # Spike times for one cell`
`      8 stimulus = data_file['stimulus']`
`----> 9 stimulus -= stimulus.mean()`
`     10 stimulus /= stimulus.std()`
 `    11 time = np.arange(stimulus.shape[0]) * data_file['stimulus'].attrs.get('frame-rate')`

`AttributeError: 'Dataset' object has no attribute 'mean'`

I installed h5py, version 2.6.0.
