### 2016-12-19 ebd0179
Adds preliminary python2.7+ support. Passes all tests. Updates setup.py and quickstart.rst to (1) fix small bug in demo code, (2) include note about HDF5 requirement for running demo, and (3) clarifies that Python 2.7 and 3+ are both supported.
