+++ setup.py
@@ -27,7 +27,8 @@ setup(name='pyret',
               'Topic :: Scientific/Engineering :: Visualization',
               'License :: OSI Approved :: MIT License',
               'Operating System :: OS Independent',
-              'Programming Language :: Python :: 3 :: Only'
+              'Programming Language :: Python :: 3',
+              'Programming Language :: Python :: 2.7'
           ],
       packages=find_packages(),
       install_requires=[
