### 2016-12-19 af2439d
Adds note about directory containing tutorial data used in demo in quickstart.rst.
