### 2016-12-19 baccuslab/pyret#86 @bnaecker
The demo does assume the working directory is `pyret/docs`. I've make that clearer in a new commit that will be merged into master soon. This also fixes the `AttributeError` you point out above.
