### 2016-12-19 baccuslab/pyret#85 @bnaecker
We're not sure why the issue with scikit-learn is happening. That package is a bit strange (for example, both it uses both `scikit-learn` and `sklearn` as package names in pip), but we'll check it out and get it fixed.

I've made the fact that the demo requires `h5py` explicit in the quickstart. However, it is not a true requirement for `pyret` itself, so we don't want to list it as such. We used it in the demo both for convenience (we use HDF5 format a lot) and because we wanted to show the fact that `pyret` will work with NumPy array types and HDF5 datasets equally well.
