### 2016-12-19 baccuslab/pyret#83 @bnaecker
Preliminary Python 2.7 support has been added in a branch I just pushed. I've assigned Niru to review the code, and we'll merge it into master when we're convinced of its correctness.
