### 2016-12-19 baccuslab/pyret#85 @kdesimone
In that case, I'd recommend that you use the Python's standard library to load data.  There are a couple options for doing this, including pickle or maybe just plain old text files. For my own purposes, I do not store any actual brain measurements for demos/testings. Rather, I synthesize the data using custom methods as needed.

This also brings up the issue where to store the data—what happens if someone installs pyret through pip?  They could import the data somehow?  Or would they need to know where pyret is installed on their system in order navigate to/specify the pyret/doc location?
