+++ .travis.yml
@@ -1,6 +1,9 @@
 language: python
 python:
+  - "2.7"
+  - "3.4"
   - "3.5"
+  - "3.6"
 install:
   - sudo apt-get -qq update
   - if [[ "$TRAVIS_PYTHON_VERSION" == "2.7" ]]; then
