### 2016-12-23 c9d86eb
Removes 3.6, adds 3.3 to travis.yml
