+++ docs/install.rst
@@ -14,7 +14,7 @@ The fastest way to install is by grabbing the code from Github:
     $ python setup.py install
 
 
-Pyret supports Python2.7+ and Python3.0+.
+Pyret supports Python2.7 and Python3.4+.
 
 Dependencies
 ------------
