### 2016-12-24 8e85a67
PR for Python 2.7 support (#87)
* Changes filter_length argument to nsamples_before, adds nsamples_after as a kwarg to filtertools

* Adds acausal functionality to filtertools.{revcorr,linear_response}.
Renames filtertools.linear_prediction -> filtertools.linear_response.
Adds acausal functionality to stimulustools.slicestim. Improves
docstrings for most of filtertools module.

* Adds tests for the acausal functionality included in previous commit.
Fixes off-by-one errors in that code. Updates some docs with more
information about the nbefore/nafter concept.

* Adds note to quickstart about re
[…]
