+++ tests/test_filtertools.py
@@ -38,7 +38,26 @@ def test_sta():
     tmp /= len(spikes)
 
     assert np.allclose(tmp, sta)
-    assert np.allclose(tax, np.arange(filter_length))
+    assert np.allclose(tax, np.arange(-filter_length + 1, 1))
+
+def test_sta_acausal():
+    """Test computing a spike-triggered average with points before and
+    after the time of the spike.
+    """
+    np.random.seed(0)
+    time = np.arange(100)
+    spikes = np.array((0, 30, 70))
+    stimulus = np.random.randn(100,)
+    nbefore, nafter = 5, 2
+
+    sta, tax = flt.sta(time, stimulus, spikes, nbefore, nafter)
+    tmp = np.zeros(sta.shape)
+    for ix in spikes[1:]: # Should ignore first spike, comes before filter_length frames
+        tmp += stimulus[ix - nbefore : ix + nafter]
+    tmp /= len(spikes)
+
+    assert np.allclose(tmp, sta)
+    assert np.allclose(tax, np.arange(-nbefore + 1, nafter + 1))
 
 def test_empty_sta():
     """Test that an empty with no spikes returns an array of nans"""
@@ -162,7 +181,7 @@ def test_normalize_spatial():
     nx, ny = 10, 10
     true_filter = utils.create_spatiotemporal_filter(nx, ny, filter_length)[1]
     noise_std = 0.01
-    noisy_filter = -(true_filter + 1.0 + np.random.randn(*true_filter.shape) * 0.01)
+    noisy_filter = true_filter + 1.0 + np.random.randn(*true_filter.shape) * 0.01
 
     normalized = flt.normalize_spatial(noisy_filter)
     normalized /= np.linalg.norm(normalized)
@@ -180,28 +199,42 @@ def test_rfsize():
     assert np.allclose(xsize, 3., 0.1) # 1 SD is about 3 units
     assert np.allclose(ysize, 3., 0.1)
 
-def test_linear_prediction_1d():
-    """Test method for computing linear prediction from a
+
+def test_linear_response_1d():
+    """Test method for computing linear response from a
     filter to a one-dimensional stimulus.
     """
     np.random.seed(0)
     filt = np.random.randn(100,)
     stim = np.random.randn(1000,)
-    pred = flt.linear_prediction(filt, stim)
+    pred = flt.linear_response(filt, stim)
 
     sl = slicestim(stim, filt.shape[0])
     assert np.allclose(sl.dot(filt), pred)
 
 
-def test_linear_prediction_nd():
-    """Test method for computing linear prediction from a
+def test_linear_response_acausal():
+    """Test computing a linear response from a filter to a 1D stimulus,
+    including acausal portions of the stimulus.
+    """
+    np.random.seed(0)
+    filt = np.random.randn(100,)
+    stim = np.random.randn(1000,)
+    pred = flt.linear_response(filt, stim, 10)
+
+    sl = slicestim(stim, filt.shape[0] - 10, 10)
+    assert np.allclose(sl.dot(filt), pred)
+
+
+def test_linear_response_nd():
+    """Test method for computing linear response from a
     filter to a multi-dimensional stimulus.
     """
     np.random.seed(0)
     for ndim in range(2, 4):
         filt = np.random.randn(100, *((10,) * ndim))
         stim = np.random.randn(1000, *((10,) * ndim))
-        pred = flt.linear_prediction(filt, stim)
+        pred = flt.linear_response(filt, stim)
 
         sl = slicestim(stim, filt.shape[0])
         tmp = np.zeros(sl.shape[0])
@@ -210,21 +243,24 @@ def test_linear_prediction_nd():
 
         assert np.allclose(tmp, pred)
 
[…]
