+++ .travis.yml
@@ -1,5 +1,7 @@
 language: python
 python:
+  - "2.7"
+  - "3.4"
   - "3.5"
 install:
   - sudo apt-get -qq update

[…]
