### 2016-12-24 4349c53
Bumps version to 0.5.1
