+++ pyret/__init__.py
@@ -24,6 +24,6 @@ __all__ = [
     'filtertools',
     ]
 
-__version__ = '0.5.0'
+__version__ = '0.5.1'
 
 from pyret import *
