+++ docs/conf.py
@@ -65,7 +65,7 @@ copyright = u'2014-2016, Benjamin Naecker, Niru Maheswaranathan'
 # The short X.Y version.
 version = '0.5'
 # The full version, including alpha/beta/rc tags.
-release = '0.5.0'
+release = '0.5.1'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
