### 2016-12-24 baccuslab/pyret#83 @nirum
Hi Kevin!

Just merged in python 2.7 support.

thanks,
Niru
