### 2016-12-24 37e6f2b
Updates requirements in setup.py, bumps version to 0.5.2
