### 2016-12-24 ee6a036
Removes pyret import from setup.py
