### 2017-01-05 e5b249c
Adds .npz file containing tutorial data, so that it can be run without install h5py. Adds clarity to quickstart.rst about how to download data in each file format from GH repo, and how to load into Python. Adds sub-section titles to demo in quickstart.rst.
