### 2017-01-05 0c9248f
adds instructions for using curl if wget is not available. fixes error in quickstart.
