### 2017-01-05 baccuslab/pyret#85 @bnaecker
We've addressed your issue with the data files as follows.

The main repo contains data in both HDF5 and NumPy's standard `.npz` file formats. We've then included in the demo instructions for downloading these files directly from the GH repo and loading the data into Python, making clear the steps and requirements for each method.

We feel that this makes the data very easy to get and use, without resorting to including the data files themselves in the PyPI package or methods for simulating it. Users just need to download the single file they'd like to use, and follow the demo without knowing any of the internals of how or where `pyret` is stored on their machine.
