### 2017-01-06 baccuslab/pyret#88 @kdesimone
JOSS review: demo errors
Hi,

I found a few small errors in the demo that throws errors ... in the third case, pyret.filtertools doesn't have a method named linear_prediction, but it does have a method named linear_response.  I guess this is what's meant?

1.
`spikes, stimulus, frame_rate = arrays['spikes], arrays['stimulus'].astype(np.float64), arrays['frame_rate'][0]`
should read
`spikes, stimulus, frame_rate = arrays['spikes'], arrays['stimulus'].astype(np.float64), arrays['frame_rate'][0]`

2.
`filter_length = int(filter_length_second / frame_rate)`
should read
`filter_length = int(filter_length_seconds / frame_rate)`

3.
`pred = pyret.filtertools.linear_prediction(sta, stimulus)`
should read
`pred = pyret.filtertools.linear_response(sta, stimulus)`
