### 2017-01-06 2cd579f
fixed name errors in quickstart.rst
