### 2017-01-06 baccuslab/pyret#88 @bnaecker
Thanks for pointing those out, all fixed in 2cd579f.
