### 2017-01-06 #137 @kdesimone
The authors have satisfactorily addressed (and closed) all the issues, which I've linked below.

https://github.com/baccuslab/pyret/issues/88
https://github.com/baccuslab/pyret/issues/86
https://github.com/baccuslab/pyret/issues/85
https://github.com/baccuslab/pyret/issues/83
