### 2017-01-06 #137 @arokem
@arfon: what's the next step here?

IIUC, @nirum and colleagues need to create a DOI for the package that incorporates the fixes from the review. Is that correct?
