### 2017-01-06 #137 @arfon
> IIUC, @nirum and colleagues need to create a DOI for the package that incorporates the fixes from the review. Is that correct?

Yes, that's correct. @nirum - please create the archive and update this thread with the DOI so we can accept this submission.
