### 2017-01-06 3bcb8c6
Bumps version to 0.5.4 (JOSS version)
