+++ pyret/__init__.py
@@ -24,6 +24,6 @@ __all__ = [
     'filtertools',
     ]
 
-__version__ = '0.5.3'
+__version__ = '0.5.4'
 
 from pyret import *
