### 2017-01-06 #137 @nirum
Thanks @kdesimone and @arokem!

I've uploaded the package using zenodo, here is the DOI:
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.232521.svg)](https://doi.org/10.5281/zenodo.232521)
