### 2017-01-06 a2b8a28
Updating paper structure and references for JOSS submission.
